requirements.txt
fastapi==0.121.3
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langdetect==1.0.9
modules==1.0.0
peft==0.17.1
PyYAML==6.0.3
transformers==4.57.1
psycopg2-binary==2.9.11
pydantic==2.12.4
PyJWT==2.10.1
python-dotenv==1.2.1
bcrypt==5.0.0
bitsandbytes==0.48.1
uvicorn==0.38.0


